move AnsiCode.java to a more general location (same as VirtualConsole.java)
