+++ src/gwt/test/org/rstudio/core/client/AnsiCodeTests.java
@@ -12,8 +12,9 @@
  * AGPL (http://www.gnu.org/licenses/agpl-3.0.txt) for more details.
  *
  */
-package org.rstudio.studio.client.workbench.views.terminal;
+package org.rstudio.core.client;
 
+import org.rstudio.core.client.AnsiCode;
 import org.rstudio.core.client.regex.Match;
 import org.rstudio.core.client.regex.Pattern;
